semantic-segmentation/requirements.txt
tensorflow-gpu==1.8
scikit-image
numpy
pillow
scipy
